+++ CHANGELOG.md
@@ -13,10 +13,9 @@
 - `-f`|`--force` option to `vclean` for usage alongside `-H`|`--remove-hubs` to remove hubs even if they are being used by running machines
 - `--delay` option to `ltest` to specify how long to wait for the lab to settle before running test scripts
 - `--show-boot-log` option to `vstart` to display the boot log (replaced the `-v`|`--verbose` option, see [Other Modified](#other-modified))
+- `--version` option to install-netkit-jh.sh
 - Ability to affect a specific user for `-T`|`--remove-tunnels` in `vclean`
-- `coreutils` as a required package to run Netkit (for `sha256sum` and `stdbuf`)
-- `util-linux` as a required package to run Netkit (for `getopt` and maybe other commands)
-- `lsof` as a required package to run Netkit (for `lsof`)
+- `bash`, `binutils`, `coreutils`, `iproute2`, `lsof`, and `util-linux` as dependencies
 
 ## Modified
 - Netkit's default configuration is stored in netkit.conf.default. This should not be modified and can be overridden with netkit.conf